+++ autonity/solidity/test/protocol.js
@@ -619,69 +619,6 @@ contract('Protocol', function (accounts) {
       assert.equal(newValInfo.liquidSupply, oldValInfo.liquidSupply, "liquidSupply changed");
     });
 
-    it('does not trigger fairness issue (unbondingStake > 0 and delegatedStake > 0)', async function () {
-      // fairness issue is triggered when delegatedStake or unbondingStake becomes 0 from positive due to slashing
-      // it can happen due to slashing rate = 100%
-      // it should not happen for slashing amount < totalStake
-      let config = JSON.parse(JSON.stringify(accountabilityConfig));
-      // modifying config so we get slashingAmount = totalStake - 1, the highest slash possible without triggering fairness issue
-      const expectedBondedStake = parseInt(config.slashingRatePrecision);
-      const expectedSlash = expectedBondedStake - 1;
-      config.collusionFactor = expectedSlash - parseInt(config.baseSlashingRateMid);
-      accountability = await AccountabilityTest.new(autonity.address, config, {from: deployer});
-      await autonity.setAccountabilityContract(accountability.address, {from:operator});
-
-      const tokenUnbondFactor = [1/10, 9/10, 1/100, 99/100, 1/1000, 999/1000, 1/10000000, 9999999/10000000];
-      const delegator = accounts[9];
-      const balance = (await autonity.balanceOf(delegator)).toNumber();
-      let validatorAddresses = [];
-      for (let i = 0; i < Math.min(validators.length, tokenUnbondFactor.length); i++) {
-        validatorAddresses.push(validators[i].nodeAddress);
-      }
-      
-      while (tokenUnbondFactor.length > validatorAddresses.length) {
-        const treasury = accounts[8];
-        const privateKey = utils.randomPrivateKey();
-        const validatorNodeAddress = await utils.registerValidator(autonity, privateKey, treasury);
-        validatorAddresses.push(validatorNodeAddress);
-      }
-
-      let tokenMinted = []
-      for (let iter = 0; iter < validatorAddresses.length; iter++) {
-        let validator = validatorAddresses[iter];
-        let validatorInfo = await autonity.getValidator(validator);
-        let bondedStake = parseInt(validatorInfo.bondedStake);
-        // non-self bond to check fairness issue
-        const tokenMint = expectedBondedStake - bondedStake;
-        tokenMinted.push(tokenMint);
-        await autonity.mint(delegator, tokenMint, {from: operator});
-        await autonity.bond(validator, tokenMint, {from: delegator});
-      }
-      // let bonding apply
-      await utils.endEpoch(autonity, operator, deployer);
-
-      for (let iter = 0; iter < validatorAddresses.length; iter++) {
-        const validator = validatorAddresses[iter];
-        let tokenUnBond = Math.max(1, Math.floor(tokenMinted[iter]*tokenUnbondFactor[iter]));
-        await autonity.unbond(validator, tokenUnBond, {from: delegator});
-      }
-      // let unbonding apply and unbondingStake create
-      await utils.endEpoch(autonity, operator, deployer);
-
-      for (let iter = 0; iter < validatorAddresses.length; iter++) {
-        const validator = validatorAddresses[iter];
-        let {txEvent, _} = await slash(config, accountability, 1, validator, validator);
-        // checking if highest possible slashing can be done without triggering fairness issue
-        // cannot slash (totalStake - 1) because both delegated and unbonding slash is floored
-        assert.equal(txEvent.amount.toNumber(), expectedSlash-1, "highest slash did not happen");
-        let validatorInfo = await autonity.getValidator(validator);
-        assert.equal(validatorInfo.state, utils.ValidatorState.jailed, "validator not jailed");
-        assert(parseInt(validatorInfo.bondedStake) > 0 && parseInt(validatorInfo.unbondingStake) > 0, "fairness issue triggered");
-      }
-      await utils.mineTillUnbondingRelease(autonity, operator, deployer);
-      assert.equal((await autonity.balanceOf(delegator)).toNumber(), balance, "unbonding released");
-    });
-
     it('kills validator for 100% slash', async function () {
       // non-self-bond and self-bond
       const validatorAddresses = [validators[0].nodeAddress, validators[1].nodeAddress];
